Player body2

+++ Assets/Scripts/Core/Player/PlayerHit.cs
@@ -5,7 +5,7 @@ namespace Core.Player
 {
     public class PlayerHit : Hittable
     {
-        public override void Hit(float damage = 1f)
+        public override void Hit(int damage = 1)
         {
             Player.Instance.ReduceDamage(gameObject.name, damage);
         }
